Fixes for tags, volume lookups and config loading
Newer aws-sdk now detects and throws an exception when tags use
"key" (a string) instead of :key (a symbol). Re-worked the tags
class with new methods:
- apitags -> provides array of hashes with key: and value:
  symbols
- cnftags -> provides array of hashes for CloudFormation templates,
  using "Key" and "Value"
Updated calls to use the correct methods.

When looking up existing volumes, we were always filtering on
attachment.device being one of /dev/sdf or /dev/xvdf - which fails
when status == available. This caused automatic volume reattachment
to fail, so removed it. NOTE: this will break detaching volumes.

Newer versions of ruby no longer use "Fixnum", so had to add
"Integer" to the list of allowed classes when loading config.