Make cloudformation lookups more robust
Instead of requiring outputs for everything that can be looked up,
the library now uses the resources associated with a stack. All
Occurrences of "output" or "outputs" in lookups have been replaced
with "resource" and "resources".

The resource specifier (formerly output specifier) previously ended
with :<resourcename>, but now takes an optional type; i.e.
:<resourcename>=<resourcetype>. E.g.
"SecurityGroups:BastionSG=AWS::EC2::SecurityGroup". If the supplied
type doesn't match the type of the named resource (e.g. BastionSG),
the library raises an exception.

Previously when lookups were output based, additional properties of
the resource could be returned if the Output was generated from
Fn::GetAtt. Since lookups now use resources, the syntax was expanded
to allow for :<resourcename>.<property>(=<resourcetype). In this
case, the resourcetype needs to match the type of resourcename, not
the property.