Updates for generating repositories param, cleanup
The next devel step will be the addition of generated parameters
for Creator, Repositories, and CFNS3Prefix. The commit hash and/or
current version of ruby-awstools will appear in the repositories
tag, so the CLI scripts have been updated for use as symlinks to
the active repository.

The environment variable for using a different file for
cloudconfig.yaml wasn't used, so it's been removed for the time
being.